implement update score script
